--- models/contacts copy.json
@@ -1,53 +1,0 @@
-[
-    {
-        "name": "Allen Raymond",
-        "email": "nulla.ante@vestibul.co.uk",
-        "phone": "(992) 914-3792"
-    },
-    {
-        "name": "Chaim Lewis",
-        "email": "dui.in@egetlacus.ca",
-        "phone": "(294) 840-6685"
-    },
-    {
-        "name": "Kennedy Lane",
-        "email": "mattis.Cras@nonenimMauris.net",
-        "phone": "(542) 451-7038"
-    },
-    {
-        "name": "Wylie Pope",
-        "email": "est@utquamvel.net",
-        "phone": "(692) 802-2949"
-    },
-    {
-        "name": "Cyrus Jackson",
-        "email": "nibh@semsempererat.com",
-        "phone": "(501) 472-5218"
-    },
-    {
-        "id": "6",
-        "name": "Abbot Franks",
-        "email": "scelerisque@magnis.org",
-        "phone": "(186) 568-3720"
-    },
-    {
-        "name": "Reuben Henry",
-        "email": "pharetra.ut@dictum.co.uk",
-        "phone": "(715) 598-5792"
-    },
-    {
-        "name": "Simon Morton",
-        "email": "dui.Fusce.diam@Donec.com",
-        "phone": "(233) 738-2360"
-    },
-    {
-        "name": "Thomas Lucas",
-        "email": "nec@Nulla.com",
-        "phone": "(704) 398-7993"
-    },
-    {
-        "name": "Alec Howard",
-        "email": "Donec.elementum@scelerisquescelerisquedui.net",
-        "phone": "(748) 206-2688"
-    }
-]
