+++ package.json
@@ -3,7 +3,8 @@
     "version": "0.0.0",
     "private": true,
     "scripts": {
-        "start:dev": "nodemon ./server.js",
+        "start": "cross-env NODE_ENV=production node ./server.js",
+        "start:dev": "cross-env NODE_ENV=development nodemon ./server.js",
         "lint": "eslint **/*.js",
         "lint:fix": "eslint --fix **/*.js"
     },
